Move `experimental` module to `io.spatial`
