+++ apis/python/tests/test_export_multiscale_image.py
@@ -7,7 +7,7 @@ import somacore
 
 import tiledbsoma as soma
 
-soma_outgest = pytest.importorskip("tiledbsoma.experimental.outgest")
+soma_outgest = pytest.importorskip("tiledbsoma.io.spatial.outgest")
 sd = pytest.importorskip("spatialdata")
